feat: prepare fiat Credits purchase
